refactor: move buildDir and outputDir helpers to cli package and drop fs-extra dependency
