+++ packages/cli/package.json
@@ -20,7 +20,6 @@
     "@sdeverywhere/compile": "^0.7.0",
     "bufx": "^1.0.5",
     "byline": "^5.0.0",
-    "fs-extra": "^9.0.0",
     "ramda": "^0.27.0",
     "shelljs": "^0.8.3",
     "yargs": "^17.2.1"
