+++ packages/cli/src/sde-log.js
@@ -1,4 +1,4 @@
-import fs from 'fs-extra'
+import fs from 'fs'
 import path from 'path'
 import R from 'ramda'
 import B from 'bufx'
