+++ packages/cli/src/sde-test.js
@@ -1,10 +1,9 @@
 import path from 'path'
 
-import { outputDir } from './Helpers.js'
 import { run } from './sde-run.js'
 import { log } from './sde-log.js'
 import { compare } from './sde-compare.js'
-import { modelPathProps } from './utils.js'
+import { modelPathProps, outputDir } from './utils.js'
 
 export let command = 'test [options] <model>'
 export let describe = 'build the model, run it, process the log, and compare to Vensim data'
