+++ packages/compile/package.json
@@ -20,7 +20,6 @@
     "byline": "^5.0.0",
     "csv-parse": "^4.15.4",
     "digraph-sort": "^0.2.1",
-    "fs-extra": "^9.0.0",
     "js-yaml": "^3.13.1",
     "ramda": "^0.27.0",
     "split-string": "^6.0.0",
